+++ web/src/app/api/chatCompletion/route.ts
@@ -1,4 +1,4 @@
-import chatCompletionHandler from "@/src/ee/features/playground/server/chatCompletionHandler";
+import chatCompletionHandler from "@/src/features/playground/server/chatCompletionHandler";
 
 export const dynamic = "force-dynamic";
 export const maxDuration = 120;
