+++ web/src/features/evals/components/evaluation-prompt-preview.tsx
@@ -1,7 +1,7 @@
 import { ItemBadge } from "@/src/components/ItemBadge";
 import { Skeleton } from "@/src/components/ui/skeleton";
-import { useExtractVariables } from "@/src/ee/features/evals/hooks/useExtractVariables";
-import { type VariableMapping } from "@/src/ee/features/evals/utils/evaluator-form-utils";
+import { useExtractVariables } from "@/src/features/evals/hooks/useExtractVariables";
+import { type VariableMapping } from "@/src/features/evals/utils/evaluator-form-utils";
 import { cn } from "@/src/utils/tailwind";
 import { type RouterOutput } from "@/src/utils/types";
 import { type EvalTemplate } from "@langfuse/shared";
