+++ web/src/features/evals/hooks/useEvalConfigMappingData.ts
@@ -1,4 +1,4 @@
-import { type EvalFormType } from "@/src/ee/features/evals/utils/evaluator-form-utils";
+import { type EvalFormType } from "@/src/features/evals/utils/evaluator-form-utils";
 import { api } from "@/src/utils/api";
 import { type ObservationType } from "@langfuse/shared";
 import { type UseFormReturn } from "react-hook-form";
