+++ web/src/features/evals/pages/new-template.tsx
@@ -1,5 +1,5 @@
 import Page from "@/src/components/layouts/page";
-import { EvalTemplateForm } from "@/src/ee/features/evals/components/template-form";
+import { EvalTemplateForm } from "@/src/features/evals/components/template-form";
 import { useHasProjectAccess } from "@/src/features/rbac/utils/checkProjectAccess";
 import { useRouter } from "next/router";
