+++ web/src/features/experiments/hooks/useEvaluatorDefaults.ts
@@ -1,7 +1,7 @@
 import { Decimal } from "decimal.js";
 import { type EvalTemplate } from "@langfuse/shared";
 import { createDefaultVariableMappings } from "../utils/evaluatorMappingUtils";
-import { type PartialConfig } from "@/src/ee/features/evals/types";
+import { type PartialConfig } from "@/src/features/evals/types";
 
 export const CONFIG_BASE = {
   targetObject: "dataset",
