+++ web/src/features/playground/page/components/PlaygroundTools/index.tsx
@@ -1,13 +1,13 @@
 import React, { useCallback, useEffect, useState } from "react";
 
-import { usePlaygroundContext } from "@/src/ee/features/playground/page/context";
+import { usePlaygroundContext } from "@/src/features/playground/page/context";
 import { Button } from "@/src/components/ui/button";
 import { ScrollArea } from "@/src/components/ui/scroll-area";
 import { PlusIcon, PencilIcon, MinusCircle, WrenchIcon } from "lucide-react";
 import { type LlmTool } from "@prisma/client";
 import { api } from "@/src/utils/api";
 import useProjectIdFromURL from "@/src/hooks/useProjectIdFromURL";
-import { CreateOrEditLLMToolDialog } from "@/src/ee/features/playground/page/components/CreateOrEditLLMToolDialog";
+import { CreateOrEditLLMToolDialog } from "@/src/features/playground/page/components/CreateOrEditLLMToolDialog";
 import {
   Command,
   CommandEmpty,
@@ -23,7 +23,7 @@ import {
   PopoverTrigger,
 } from "@/src/components/ui/popover";
 
-import { type PlaygroundTool } from "@/src/ee/features/playground/page/types";
+import { type PlaygroundTool } from "@/src/features/playground/page/types";
 
 export const PlaygroundTools = () => {
   const { tools, setTools } = usePlaygroundContext();
