+++ web/src/pages/project/[projectId]/annotation-queues.tsx
@@ -1,3 +1,3 @@
-import AnnotationQueues from "@/src/ee/features/annotation-queues/pages/AnnotationQueues";
+import AnnotationQueues from "@/src/features/annotation-queues/pages/AnnotationQueues";
 
 export default AnnotationQueues;
