+++ web/src/pages/project/[projectId]/annotation-queues/[queueId]/index.tsx
@@ -1,3 +1,3 @@
-import QueueItems from "@/src/ee/features/annotation-queues/pages/AnnotationQueueItems";
+import QueueItems from "@/src/features/annotation-queues/pages/AnnotationQueueItems";
 
 export default QueueItems;
