+++ web/src/pages/project/[projectId]/evals/[evaluatorId].tsx
@@ -1,3 +1,3 @@
-import { EvaluatorDetail } from "@/src/ee/features/evals/components/evaluator-detail";
+import { EvaluatorDetail } from "@/src/features/evals/components/evaluator-detail";
 
 export default EvaluatorDetail;
