+++ web/src/pages/project/[projectId]/evals/default-model.tsx
@@ -1,1 +1,1 @@
-export { default as default } from "@/src/ee/features/evals/pages/default-evaluation-model";
+export { default as default } from "@/src/features/evals/pages/default-evaluation-model";
