+++ web/src/pages/project/[projectId]/evals/index.tsx
@@ -1,1 +1,1 @@
-export { default as default } from "@/src/ee/features/evals/pages/evaluators";
+export { default as default } from "@/src/features/evals/pages/evaluators";
