+++ web/src/pages/project/[projectId]/evals/new.tsx
@@ -1,1 +1,1 @@
-export { default as default } from "@/src/ee/features/evals/pages/new-evaluator";
+export { default as default } from "@/src/features/evals/pages/new-evaluator";
