+++ web/src/pages/project/[projectId]/evals/templates/new.tsx
@@ -1,1 +1,1 @@
-export { default as default } from "@/src/ee/features/evals/pages/new-template";
+export { default as default } from "@/src/features/evals/pages/new-template";
